requirements.txt
# Core
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0

# UI
streamlit>=1.28.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
faker>=22.0.0

# LLM Integration
groq>=0.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: Local LLM
# ollama>=0.1.0
